feat(payeeEthBalance): change reducer to module
