+++ src/containers/transfer/components/TokenTransferForm.js
@@ -14,7 +14,8 @@ import Button from "components/augmint-ui/button";
 import { Form, Validations, Normalizations, Parsers } from "components/BaseComponents";
 import { getTransferFee } from "modules/ethereum/transferTransactions";
 import { transferToken, TOKEN_TRANSFER_SUCCESS } from "modules/reducers/augmintToken";
-import { refreshPayeesEthBalance, PAYEE_ETH_BALANCE_RECEIVED } from "modules/reducers/payeeEthBalance";
+// import { refreshPayeesEthBalance, PAYEE_ETH_BALANCE_RECEIVED } from "modules/reducers/payeeEthBalance";
+import { getPayeesEthBalance } from "modules/payeeEthBalance";
 import { TransferFeeToolTip } from "./AccountToolTips";
 import theme from "styles/theme";
 
@@ -85,15 +86,15 @@ class TokenTransferForm extends React.Component {
         const tokenAmount = parseFloat(values.tokenAmount);
         // console.log("payee: ", values.payee);
 
-        const payeeEthBalance = await store.dispatch(refreshPayeesEthBalance(values.payee));
+        // const payeeEthBalance = await store.dispatch(refreshPayeesEthBalance(values.payee));
+        const payeeEthBalance = await getPayeesEthBalance(values.payee);
+        console.log("balance: ", payeeEthBalance);
 
-        // console.log("balance: ", payeeEthBalance);
-
-        if (payeeEthBalance.type !== PAYEE_ETH_BALANCE_RECEIVED) {
-            throw new SubmissionError({
-                _error: payeeEthBalance.error
-            });
-        }
+        // if (payeeEthBalance.type !== PAYEE_ETH_BALANCE_RECEIVED) {
+        //     throw new SubmissionError({
+        //         _error: payeeEthBalance.error
+        //     });
+        // }
 
         const res = await store.dispatch(
             transferToken({
@@ -107,7 +108,7 @@ class TokenTransferForm extends React.Component {
                 _error: res.error
             });
         } else {
-            if (payeeEthBalance.ethBalance !== 0) {
+            if (!payeeEthBalance.error && payeeEthBalance.ethBalance !== 0) {
                 this.setPayeeAddress(values.payee);
             }
